intelli_plug: disable poll waiting to get re-activated
intelli_plug queue gets called when the screen is back on.

This saves a few CPU cycles during screen-off but introduces a new bug
which users having to turn the screen off and back on if they manually
wrote a new value to intelli_plug_active.

Be cautious if you don't have intelli_plug activated by default on the
source-code side.

Signed-off-by: Park Ju Hyung <qkrwngud825@gmail.com>